Remove assembler_riscv_c.hpp as discussions
